long을 integer로 일괄 변경

+++ BurgerStack_Project/src/main/java/com/kh/burgerstack/closing/StoreClosing.java
@@ -15,9 +15,9 @@ import lombok.ToString;
 @Setter
 @ToString
 public class StoreClosing {
-    private Long storeClosingId;
+    private Integer storeClosingId;
     private LocalDate businessDate;
     private String closingMemo;
     private LocalDateTime closedAt;
-    private Long storeId;
+    private Integer storeId;
 }
