+++ BurgerStack_Project/src/main/java/com/kh/burgerstack/inquiry/InquiryFile.java
@@ -14,12 +14,12 @@ import lombok.ToString;
 @Setter
 @ToString
 public class InquiryFile {
-    private Long inquiryFileId;
+    private Integer inquiryFileId;
     private String originalName;
     private String storedName;
     private String storagePath;
     private String attachTarget;
     private LocalDateTime createdAt;
     private LocalDateTime deletedAt;
-    private Long inquiryId;
+    private Integer inquiryId;
 }
