+++ BurgerStack_Project/src/main/java/com/kh/burgerstack/store/StoreMapper.java
@@ -4,5 +4,5 @@ import org.apache.ibatis.annotations.Mapper;
 
 @Mapper
 public interface StoreMapper {
-    public Long findStoreIdByOwnerUserNo(Long ownerUserNo);
+    public Integer findStoreIdByOwnerUserNo(Integer ownerUserNo);
 }
